Moved upload command into its own module.
