+++ src/platform/testing/e2e/cypress/support/index.js
@@ -1,6 +1,6 @@
 import '@testing-library/cypress/add-commands';
 
-import './commands';
+import './upload';
 
 // Workaround to allow Cypress to intercept requests made with the Fetch API.
 // This forces fetch to fall back to the polyfill, which can get intercepted.
