+++ src/platform/testing/e2e/cypress/support/upload.js
